--- a/example/.gitignore
+++ b/example/.gitignore
@@ -1,2 +1,1 @@
 data
-any.py