--- a/server/src/data/repository/cache_repo.rs
+++ b/server/src/data/repository/cache_repo.rs
@@ -9,6 +9,7 @@ const REDIS_KEY_CACHE_CONFIG: &str = "request_cache_config"; // hash
 
 #[async_trait]
 pub trait CacheRepo {
+    fn enabled(&self) -> bool;
     async fn get(&self, key: String) -> Result<Cache, String>;
     async fn set(&self, key: String, data: Cache) -> Result<(), String>;
     async fn get_configs(&self) -> Result<Vec<CacheConfig>, String>;
@@ -17,18 +18,23 @@ pub trait CacheRepo {
     async fn remove_config(&self, key: String) -> Result<(), String>;
 }
 
-pub struct CacheRepoImpl {
+// Cache with redis implementation
+pub struct CacheRepoRedisImpl {
     connection: MultiplexedConnection
 }
 
-impl CacheRepoImpl {
+impl CacheRepoRedisImpl {
     pub fn new(connection: MultiplexedConnection) -> Self {
-        CacheRepoImpl { connection }
+        CacheRepoRedisImpl { connection }
     }
 }
 
 #[async_trait]
-impl CacheRepo for CacheRepoImpl {
+impl CacheRepo for CacheRepoRedisImpl {
+    fn enabled(&self) -> bool {
+        true
+    }
+
     async fn get(&self, key: String) -> Result<Cache, String> {
         let data: Vec<u8> = self.connection.clone().hget(REDIS_KEY_CACHE, key.clone()).await
             .map_err(|e| format!("Error getting cache {}: {}", key, e))?;
@@ -86,5 +92,54 @@ impl CacheRepo for CacheRepoImpl {
             .map_err(|e| format!("Error unsetting cache config {}: {}", key, e))?;
 
         Ok(())
+    }
+}
+
+// cache with in process memory implementation
+// TODO: implement this
+//       might try to refactor redis config to be stored locally in file
+pub struct CacheRepoProcMemImpl {
+    // cache: Arc<Mutex<HashMap<String, Cache>>>,
+    // cache_config: Arc<Mutex<HashMap<String, CacheConfig>>>
+}
+
+impl CacheRepoProcMemImpl {
+    pub fn new() -> Self {
+        CacheRepoProcMemImpl {
+            // cache: Arc::new(Mutex::new(HashMap::new())),
+            // cache_config: Arc::new(Mutex::new(HashMap::new())),
+        }
+    }
+}
+
+#[async_trait]
+impl CacheRepo for CacheRepoProcMemImpl {
+    fn enabled(&self) -> bool {
+        // TODO: enable after implementation
+        false
+    }
+
+    async fn get(&self, key: String) -> Result<Cache, String> {
+        Err(String::from("Cache is only available for redis store."))
+    }
+
+    async fn set(&self, key: String, cache: Cache) -> Result<(), String> {
+        Err(String::from("Cache is only available for redis store."))
+    }
+
+    async fn get_configs(&self) -> Result<Vec<CacheConfig>, String> {
+        Err(String::from("Cache is only available for redis store."))
+    }
+
+    async fn get_config(&self, key: String) -> Result<CacheConfig, String> {
+        Err(String::from("Cache is only available for redis store."))
+    }
+
+    async fn set_config(&self, key: String, config: CacheConfig) -> Result<(), String> {
+        Err(String::from("Cache is only available for redis store."))
+    }
+
+    async fn remove_config(&self, key: String) -> Result<(), String> {
+        Err(String::from("Cache is only available for redis store."))
     }
 }