--- a/server/src/data/repository/client_repo.rs
+++ b/server/src/data/repository/client_repo.rs
@@ -1,7 +1,8 @@
-use std::time::SystemTime;
+use std::{collections::HashMap, sync::Arc, time::SystemTime};
 
 use async_trait::async_trait;
 use redis::{aio::MultiplexedConnection, AsyncCommands};
+use tokio::sync::Mutex;
 use common::{convert::{from_json_slice, to_json_vec}, data::dto::tunnel_client::TunnelClient};
 
 const REDIS_KEY_CLIENT: &str = "tunnel_clients";
@@ -13,18 +14,19 @@ pub trait ClientRepo {
     async fn set_dc(&self, id: String, dt: SystemTime) -> Result<(), String>;
 }
 
-pub struct ClientRepoImpl {
+// Redis implementation
+pub struct ClientRepoRedisImpl {
     connection: MultiplexedConnection
 }
 
-impl ClientRepoImpl {
+impl ClientRepoRedisImpl {
     pub fn new(connection: MultiplexedConnection) -> Self {
-        ClientRepoImpl { connection }
+        ClientRepoRedisImpl { connection }
     }
 }
 
 #[async_trait]
-impl ClientRepo for ClientRepoImpl {
+impl ClientRepo for ClientRepoRedisImpl {
     async fn get(&self, id: String) -> Result<TunnelClient, String> {
         let data: Vec<u8> = self.connection.clone().hget(REDIS_KEY_CLIENT, id.clone()).await
             .map_err(|e| format!("Error getting client {}: {}", id, e))?;
@@ -47,5 +49,38 @@ impl ClientRepo for ClientRepoImpl {
         let mut curr_data = self.get(id).await?;
         curr_data.conn_dc_at = Option::from(dt);
         self.create(curr_data).await
+    }
+}
+
+// In process memory implementation
+pub struct ClientRepoProcMemImpl {
+    data: Arc<Mutex<HashMap<String, TunnelClient>>>
+}
+
+impl ClientRepoProcMemImpl {
+    pub fn new() -> Self {
+        ClientRepoProcMemImpl { data: Arc::new(Mutex::new(HashMap::new())) }
+    }
+}
+
+#[async_trait]
+impl ClientRepo for ClientRepoProcMemImpl {
+    async fn get(&self, id: String) -> Result<TunnelClient, String> {
+        if let Some(value) = self.data.lock().await.get(&id) {
+            return Ok((*value).clone());
+        }
+        Err(String::from("Error getting client: no valid client exists"))
+    }
+
+    async fn create(&self, client: TunnelClient) -> Result<(), String> {
+        let key = client.clone().id;            
+        self.data.lock().await.insert(key, client);
+        Ok(())
+    }
+
+    async fn set_dc(&self, id: String, dt: SystemTime) -> Result<(), String> {
+        let mut curr = self.get(id).await?;
+        curr.conn_dc_at = Option::from(dt);
+        self.create(curr).await
     }
 }