--- a/server/src/handler/public_handler.rs
+++ b/server/src/handler/public_handler.rs
@@ -3,6 +3,7 @@ use std::io::{Cursor, Read};
 
 use chrono::Utc;
 use common::convert::{parse_request_bytes, request_to_bytes, modify_headers_of_response_bytes};
+use common::data::dto::cache_config::CacheConfig;
 use common::net::{
     http_json_response_as_bytes,
     get_cookie_from_request,
@@ -124,25 +125,28 @@ async fn public_handler(
     let request_uri = request.uri().to_string();
     let request_method = String::from(request.method().as_str());
     let request_body = get_unique_body_as_bytes(request.clone());
-    let cache_config = cache_service.get_cache_config(client_id.clone(), request_method.clone(), path.clone()).await;
 
     info!("Public Request: [{}] [{}] {}", request_id.clone(), request_method.clone(), request_uri.clone());
 
-    match cache_config.clone() {
-        Ok(_) => {
-            match cache_service.get_cache(client_id.clone(), request_uri.clone(), request_method.clone(), request_body.clone()).await {
-                Ok(cached_response) => {
-                    info!("Public Request: {} processed.", request_id);
-        
-                    // return the cached response to public client
-                    stream.write_all(&cached_response).await.unwrap();
-        
-                    return
-                },
-                Err(msg) => { error!("Error getting cache for request {}: {}", request_id.clone(), msg) } // ignore error
-            }
-        },
-        Err(_) => {}
+    let mut cache_config: Result<CacheConfig, String> = Err(String::from("default"));
+    if cache_service.is_enabled() {
+        cache_config = cache_service.get_cache_config(client_id.clone(), request_method.clone(), path.clone()).await;
+        match cache_config {
+            Ok(_) => {
+                match cache_service.get_cache(client_id.clone(), request_uri.clone(), request_method.clone(), request_body.clone()).await {
+                    Ok(cached_response) => {
+                        info!("Public Request: {} processed.", request_id);
+            
+                        // return the cached response to public client
+                        stream.write_all(&cached_response).await.unwrap();
+            
+                        return
+                    },
+                    Err(msg) => { error!("Error getting cache for request {}: {}", request_id.clone(), msg) } // ignore error
+                }
+            },
+            Err(_) => {}
+        }
     }
 
     let public_request = PublicRequest {
@@ -186,13 +190,15 @@ async fn public_handler(
         false => None
     });
 
-    match cache_config {
-        Ok(config) => {
-            if let Err(msg) = cache_service.set_cache(client_id, request_uri, request_method, request_body, res.clone(), config).await {
-                error!("Error writing cache for request {}: {}", request_id.clone(), msg);
-            }
-        },
-        Err(_) => {}
+    if cache_service.is_enabled() {
+        match cache_config {
+            Ok(config) => {
+                if let Err(msg) = cache_service.set_cache(client_id, request_uri, request_method, request_body, res.clone(), config).await {
+                    error!("Error writing cache for request {}: {}", request_id.clone(), msg);
+                }
+            },
+            Err(_) => {}
+        }
     }
 
     info!("Public Request: {} processed.", request_id);