Background:
- A known issue when deploying on a low-resource server could lead to instability to the whole server.
- This is because of the `pull` model of the executor to check data from Redis continuously.

Checklist:
- [x] Try implementing non in-process option without Redis, but still use the same method
  - Testing on 25 * 40 (1000) requests
    With In-Process Memory (faster by 6%):
    ```
    Try 1:
    Elapsed time: 26.7615 seconds
    request_cnt: 2000
    success_cnt:  2000
    
    Try 2:
    Elapsed time: 28.2461 seconds
    request_cnt: 2000
    success_cnt:  2000
    
    Try 3:
    Elapsed time: 27.2055 seconds
    request_cnt: 2000
    success_cnt:  2000
    ```

    With Redis:
    ```
    Elapsed time: 30.1883 seconds
    request_cnt: 2000
    success_cnt:  2000
    
    Elapsed time: 28.4634 seconds
    request_cnt: 2000
    success_cnt:  2000
    
    Elapsed time: 28.6372 seconds
    request_cnt: 2000
    success_cnt:  2000
    ```
  - Resource on task manager (windows): 
    With In-Process Memory (significantly better in processor and memory):
    ![image](https://github.com/user-attachments/assets/6f04e780-696b-4ad3-add8-b42b68fd8a49)

    With Redis:
    ![image](https://github.com/user-attachments/assets/6b10d226-f511-4012-9183-7ce05c8b9afe)
    
    *Note: the compute usage only shows the percentage of the total power, which might be different in another env.

- TODO: next phase: Change `pull` to `push` (?)


Added:
- No Redis Option for better performance

Changed:
- Doc